+++ cloudflare/table_cloudflare_zone.go
@@ -15,7 +15,7 @@ func tableCloudflareZone(ctx context.Context) *plugin.Table {
 		Name:        "cloudflare_zone",
 		Description: "A Zone is a domain name along with its subdomains and other identities.",
 		List: &plugin.ListConfig{
-			Hydrate: listZone,
+			Hydrate: listZones,
 		},
 		Get: &plugin.GetConfig{
 			KeyColumns:        plugin.SingleColumn("id"),
@@ -26,13 +26,14 @@ func tableCloudflareZone(ctx context.Context) *plugin.Table {
 			// Top columns
 			{Name: "id", Type: proto.ColumnType_STRING, Description: "Zone identifier tag."},
 			{Name: "name", Type: proto.ColumnType_STRING, Description: "The domain name."},
+
 			// Other columns
 			// TODO - do we need this here {Name: "account", Type: proto.ColumnType_JSON, Description: "TODO"},
 			{Name: "betas", Type: proto.ColumnType_JSON, Description: "Beta feature flags associated with the zone."},
 			{Name: "created_on", Type: proto.ColumnType_TIMESTAMP, Description: "When the zone was created."},
 			{Name: "deactivation_reason", Type: proto.ColumnType_STRING, Description: "TODO"},
 			{Name: "development_mode", Type: proto.ColumnType_INT, Description: "The interval (in seconds) from when development mode expires (positive integer) or last expired (negative integer) for the domain. If development mode has never been enabled, this value is 0."},
-			//{Name: "dnssec", Type: proto.ColumnType_JSON, Hydrate: getZoneDNSSEC, Transform: transform.FromValue(), Description: "DNSSEC settings for the zone."},
+			{Name: "dnssec", Type: proto.ColumnType_JSON, Hydrate: getZoneDNSSEC, Transform: transform.FromValue(), Description: "DNSSEC settings for the zone."},
 			{Name: "host", Type: proto.ColumnType_JSON, Description: "TODO"},
 			{Name: "meta", Type: proto.ColumnType_JSON, Description: "Metadata associated witht the zone."},
 			{Name: "modified_on", Type: proto.ColumnType_TIMESTAMP, Description: "When the zone was last modified."},
@@ -43,7 +44,7 @@ func tableCloudflareZone(ctx context.Context) *plugin.Table {
 			{Name: "owner", Type: proto.ColumnType_JSON, Description: "Information about the user or organization that owns the zone."},
 			{Name: "paused", Type: proto.ColumnType_BOOL, Description: "Indicates if the zone is only using Cloudflare DNS services. A true value means the zone will not receive security or performance benefits."},
 			{Name: "permissions", Type: proto.ColumnType_JSON, Description: "Available permissions on the zone for the current user requesting the item."},
-			//{Name: "settings", Type: proto.ColumnType_JSON, Hydrate: getZoneSettings, Transform: transform.FromValue().Transform(settingsToStandard), Description: "Simple key value map of zone settings like advanced_ddos = on. Full settings details are in settings_src."},
+			{Name: "settings", Type: proto.ColumnType_JSON, Hydrate: getZoneSettings, Transform: transform.FromValue().Transform(settingsToStandard), Description: "Simple key value map of zone settings like advanced_ddos = on. Full settings details are in settings_src."},
 			//{Name: "settings_src", Type: proto.ColumnType_JSON, Hydrate: getZoneSettings, Transform: transform.FromValue(), Description: "Original source form of zone settings for caching, security and other features of Cloudflare."},
 			{Name: "plan", Type: proto.ColumnType_JSON, Description: "Current plan associated with the zone."},
 			{Name: "plan_pending", Type: proto.ColumnType_JSON, Description: "Pending plan change associated with the zone."},
@@ -56,13 +57,16 @@ func tableCloudflareZone(ctx context.Context) *plugin.Table {
 	}
 }
 
-func listZone(ctx context.Context, d *plugin.QueryData, _ *plugin.HydrateData) (interface{}, error) {
+func listZones(ctx context.Context, d *plugin.QueryData, _ *plugin.HydrateData) (interface{}, error) {
+	logger := plugin.Logger(ctx)
 	conn, err := connect(ctx, d)
 	if err != nil {
+		logger.Error("listZones", "connection_error", err)
 		return nil, err
 	}
 	resp, err := conn.ListZonesContext(ctx)
 	if err != nil {
+		logger.Error("listZones", "ListZonesContext api error", err)
 		return nil, err
 	}
 	for _, i := range resp.Result {
@@ -115,19 +119,6 @@ func getZoneDNSSEC(ctx context.Context, d *plugin.QueryData, h *plugin.HydrateDa
 	}
 	zone := h.Item.(cloudflare.Zone)
 	item, err := conn.ZoneDNSSECSetting(ctx, zone.ID)
-	if err != nil {
-		return nil, err
-	}
-	return item, nil
-}
-
-func getZoneUniversalSSLSettings(ctx context.Context, d *plugin.QueryData, h *plugin.HydrateData) (interface{}, error) {
-	conn, err := connect(ctx, d)
-	if err != nil {
-		return nil, err
-	}
-	zone := h.Item.(cloudflare.Zone)
-	item, err := conn.UniversalSSLSettingDetails(ctx, zone.ID)
 	if err != nil {
 		return nil, err
 	}